backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
anthropic>=0.43.0
pydantic>=2.10.0
python-dotenv==1.0.1
pillow>=11.0.0
boto3>=1.35.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
